chore(source file name): renames the source files PE-2814
