+++ lib/utils/snapshots/tx_snapshot_to_snapshot_data.dart
